+++ crates/algorithms/src/shortest_paths/mod.rs
@@ -4,7 +4,7 @@ mod astar;
 mod bellman_ford;
 mod dijkstra;
 mod floyd_warshall;
-mod k_shortest_paths;
+mod k_shortest_path_length;
 mod measure;
 mod total_ord;
 
@@ -12,6 +12,6 @@ pub use astar::astar;
 pub use bellman_ford::{bellman_ford, find_negative_cycle, Paths};
 pub use dijkstra::dijkstra;
 pub use floyd_warshall::floyd_warshall;
-pub use k_shortest_paths::k_shortest_paths;
+pub use k_shortest_path_length::k_shortest_path_length;
 pub use measure::{BoundedMeasure, FloatMeasure, Measure};
 pub use total_ord::TotalOrd;