+++ tests/quickcheck.rs
@@ -19,8 +19,8 @@ use petgraph::{
     algo::{
         bellman_ford, condensation, dijkstra, find_negative_cycle, floyd_warshall,
         greedy_feedback_arc_set, greedy_matching, is_cyclic_directed, is_cyclic_undirected,
-        is_isomorphic, is_isomorphic_matching, k_shortest_path, kosaraju_scc, maximum_matching,
-        min_spanning_tree, tarjan_scc, toposort, Matching,
+        is_isomorphic, is_isomorphic_matching, k_shortest_path_length, kosaraju_scc,
+        maximum_matching, min_spanning_tree, tarjan_scc, toposort, Matching,
     },
     data::FromElements,
     dot::{Config, Dot},